Changes to server sub adds log4perl
Watch for changes and restart if so. Does not really
work all that well and I am sure I am doing something
very wrong or maybe not. Anyway, we update the server
sub a bit and prepare for more work on Web.pm. I am
using fork here a bit, but the real idea is to get the
watcher watching in a different process so we can
push something to the server. Adds log4perl as well.